Fix eslint errors in Containers
