+++ src/Api/MoiraApiInjection.jsx
@@ -11,8 +11,11 @@ export function withMoiraApi<Config: {}>(
 ): React.AbstractComponent<$Diff<Config, { moiraApi: IMoiraApi | void }>> {
     return class extends React.Component<$Diff<Config, { moiraApi: IMoiraApi | void }>> {
         context: IMoiraApi;
+
         static displayName = `withApi(${Component.displayName || Component.name || "Component"})`;
+
         static contextType = ApiContext;
+
         render(): React.Element<typeof Component> {
             return <Component moiraApi={this.context} {...this.props} />;
         }
