--- src/Containers/NotificationListContainer.less
@@ -1,9 +1,0 @@
-@import '~styles/variables.less';
-@import '~styles/mixins.less';
-
-.actions-row {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-content: center;
-}
