+++ src/Containers/SettingsContainer.jsx
@@ -9,9 +9,10 @@ import type { Subscription } from "../Domain/Subscription";
 import { withMoiraApi } from "../Api/MoiraApiInjection";
 import Layout, { LayoutContent, LayoutTitle } from "../Components/Layout/Layout";
 import ContactList from "../Components/ContactList/ContactList";
-import SubscriptionList, { type SubscriptionInfo } from "../Components/SubscriptionList/SubscriptionList";
+import SubscriptionList, {
+    type SubscriptionInfo,
+} from "../Components/SubscriptionList/SubscriptionList";
 import type { NewContactInfo } from "../Components/NewContactModal/NewContactModal";
-import cn from "./SettingsContainer.less";
 
 type Props = ContextRouter & { moiraApi: IMoiraApi };
 type State = {
@@ -24,6 +25,7 @@ type State = {
 
 class SettingsContainer extends React.Component<Props, State> {
     props: Props;
+
     state: State = {
         config: null,
         loading: true,
@@ -36,43 +38,60 @@ class SettingsContainer extends React.Component<Props, State> {
         this.getData();
     }
 
-    normalizeContactValueForApi(contactType: string, value: string): string {
+    static normalizeContactValueForApi(contactType: string, value: string): string {
         let result = value.trim();
         if (contactType === "twilio voice" || contactType === "twilio sms") {
             if (result.length >= 11) {
                 result = result.replace(/^8/, "+7");
                 result = result.replace(/^7/, "+7");
             } else if (result.length === 10) {
-                result = "+7" + result;
+                result = `+7${result}`;
             }
             return result;
         }
         return result;
     }
 
-    normalizeContactValueForUi(contactType: string, value: string): string {
+    static normalizeContactValueForUi(contactType: string, value: string): string {
         return value;
     }
 
-    async getData() {
-        const { moiraApi } = this.props;
-        try {
-            const tags = (await moiraApi.getTagList()).list;
-            let settings = await moiraApi.getSettings();
-            const config = await moiraApi.getConfig();
-            settings = {
-                ...settings,
-                contacts: settings.contacts.map(x => ({
-                    ...x,
-                    value: this.normalizeContactValueForUi(x.type, x.value),
-                })),
-            };
-            this.setState({ settings: settings, config: config, tags: tags });
-        } catch (error) {
-            this.setState({ error: error.message });
-        } finally {
-            this.setState({ loading: false });
-        }
+    render(): React.Node {
+        const { loading, error, tags, settings, config } = this.state;
+        return (
+            <Layout loading={loading} error={error}>
+                <LayoutContent>
+                    <LayoutTitle>Notifications</LayoutTitle>
+                    {config != null && settings != null && settings.contacts != null && (
+                        <div style={{ marginBottom: 50 }}>
+                            <ContactList
+                                contactDescriptions={config.contacts}
+                                items={settings.contacts}
+                                onTestContact={this.handleTestContact}
+                                onAddContact={this.handleAddContact}
+                                onUpdateContact={this.handleUpdateContact}
+                                onRemoveContact={this.handleRemoveContact}
+                            />
+                        </div>
+                    )}
+                    {settings != null &&
+                        tags != null &&
+                        settings.subscriptions != null &&
+                        settings.contacts != null &&
+                        settings.contacts.length > 0 && (
+                            <SubscriptionList
+                                tags={tags}
+                                contacts={settings.contacts}
+                                subscriptions={settings.subscriptions}
+                                onTestSubscription={this.handleTestSubscription}
+                                onAddSubscription={this.handleAddSubscription}
+                                onRemoveSubscription={this.handleRemoveSubscription}
+                                onUpdateSubscription={this.handleUpdateSubscription}
+                            />
+                        )}
+                </LayoutContent>
+            </Layout>
+        );
     }
 
     handleTestContact = async (contact: Contact) => {
@@ -104,13 +123,16 @@ class SettingsContainer extends React.Component<Props, State> {
 
         try {
             let newContact = await moiraApi.addContact({
-                value: this.normalizeContactValueForApi(contactType, contact.value),
+                value: SettingsContainer.normalizeContactValueForApi(contactType, contact.value),
                 type: contactType,
                 user: settings.login,
             });
             newContact = {
                 ...newContact,
-                value: this.normalizeContactValueForUi(newContact.type, newContact.value),
+                value: SettingsContainer.normalizeContactValueForUi(
+                    newContact.type,
+                    newContact.value
+                ),
             };
             this.setState({
                 settings: {
@@ -135,7 +157,7 @@ class SettingsContainer extends React.Component<Props, State> {
         try {
             await moiraApi.updateContact({
                 ...contact,
-                value: this.normalizeContactValueForApi(contact.type, contact.value),
+                value: SettingsContainer.normalizeContactValueForApi(contact.type, contact.value),
             });
             const index = contacts.findIndex(x => x.id === contact.id);
             this.setState({
@@ -193,7 +215,11 @@ class SettingsContainer extends React.Component<Props, State> {
             this.setState({
                 settings: {
                     ...settings,
-                    subscriptions: [...subscriptions.slice(0, index), subscription, ...subscriptions.slice(index + 1)],
+                    subscriptions: [
+                        ...subscriptions.slice(0, index),
+                        subscription,
+                        ...subscriptions.slice(index + 1),
+                    ],
                 },
             });
         } catch (error) {
@@ -239,42 +265,25 @@ class SettingsContainer extends React.Component<Props, State> {
         }
     };
 
-    render(): React.Node {
-        const { loading, error, tags, settings, config } = this.state;
-        return (
-            <Layout loading={loading} error={error}>
-                <LayoutContent>
-                    <LayoutTitle>Notifications</LayoutTitle>
-                    {config != null && settings != null && settings.contacts != null && (
-                        <div className={cn("contact-list")}>
-                            <ContactList
-                                contactDescriptions={config.contacts}
-                                items={settings.contacts}
-                                onTestContact={this.handleTestContact}
-                                onAddContact={this.handleAddContact}
-                                onUpdateContact={this.handleUpdateContact}
-                                onRemoveContact={this.handleRemoveContact}
-                            />
-                        </div>
-                    )}
-                    {settings != null &&
-                        tags != null &&
-                        settings.subscriptions != null &&
-                        settings.contacts != null &&
-                        settings.contacts.length > 0 && (
-                            <SubscriptionList
-                                tags={tags}
-                                contacts={settings.contacts}
-                                subscriptions={settings.subscriptions}
-                                onTestSubscription={this.handleTestSubscription}
-                                onAddSubscription={this.handleAddSubscription}
-                                onRemoveSubscription={this.handleRemoveSubscription}
-                                onUpdateSubscription={this.handleUpdateSubscription}
-                            />
-                        )}
-                </LayoutContent>
-            </Layout>
-        );
+    async getData() {
+        const { moiraApi } = this.props;
+        try {
+            const tags = (await moiraApi.getTagList()).list;
+            let settings = await moiraApi.getSettings();
+            const config = await moiraApi.getConfig();
+            settings = {
+                ...settings,
+                contacts: settings.contacts.map(x => ({
+                    ...x,
+                    value: SettingsContainer.normalizeContactValueForUi(x.type, x.value),
+                })),
+            };
+            this.setState({ settings, config, tags });
+        } catch (error) {
+            this.setState({ error: error.message });
+        } finally {
+            this.setState({ loading: false });
+        }
     }
 }
