--- src/Containers/SettingsContainer.less
@@ -1,3 +1,0 @@
-.contact-list {
-    margin-bottom: 50px;
-}
