--- src/Containers/TriggerContainer.less
@@ -1,3 +1,0 @@
-.empty-details-text {
-    color: #888888;
-}
