+++ src/Containers/TriggerEditContainer.jsx
@@ -1,13 +1,13 @@
 // @flow
 import * as React from "react";
 import type { ContextRouter } from "react-router-dom";
+import { ValidationContainer } from "react-ui-validations";
+import Button from "retail-ui/components/Button";
+import TrashIcon from "@skbkontur/react-icons/Trash";
 import type { IMoiraApi } from "../Api/MoiraApi";
 import type { Trigger } from "../Domain/Trigger";
 import { getPageLink } from "../Domain/Global";
 import { withMoiraApi } from "../Api/MoiraApiInjection";
-import { ValidationContainer } from "react-ui-validations";
-import Button from "retail-ui/components/Button";
-import TrashIcon from "@skbkontur/react-icons/Trash";
 import RouterLink from "../Components/RouterLink/RouterLink";
 import Layout, { LayoutContent, LayoutTitle } from "../Components/Layout/Layout";
 import TriggerEditForm from "../Components/TriggerEditForm/TriggerEditForm";
@@ -24,15 +24,21 @@ type State = {
 };
 
 class TriggerEditContainer extends React.Component<Props, State> {
-    props: Props;
-    state: State = {
-        loading: true,
-        error: null,
-        config: null,
-        trigger: null,
-        tags: null,
-    };
-    triggerForm: ?ValidationContainer;
+    state: State;
+
+    validationContainer: { current: null | ValidationContainer };
+
+    constructor(props: Props) {
+        super(props);
+        this.state = {
+            loading: true,
+            error: null,
+            config: null,
+            trigger: null,
+            tags: null,
+        };
+        this.validationContainer = React.createRef<ValidationContainer>();
+    }
 
     componentDidMount() {
         this.getData(this.props);
@@ -43,34 +49,63 @@ class TriggerEditContainer extends React.Component<Props, State> {
         this.getData(nextProps);
     }
 
-    async getData(props: Props) {
-        const { moiraApi, match } = props;
-        const { id } = match.params;
-        if (typeof id !== "string") {
-            this.setState({ error: "Wrong trigger id", loading: false });
-            return;
-        }
-        try {
-            const trigger = await moiraApi.getTrigger(id);
-            const { list } = await moiraApi.getTagList();
-            const config = await moiraApi.getConfig();
-            this.setState({
-                trigger: trigger,
-                tags: list,
-                config: config,
-            });
-        } catch (error) {
-            this.setState({ error: error.message });
-        } finally {
-            this.setState({ loading: false });
-        }
-    }
-
-    async validateForm(): Promise<boolean> {
-        if (this.triggerForm == null) {
-            return true;
-        }
-        return await this.triggerForm.validate();
+    render(): React.Node {
+        const { loading, error, trigger, tags, config } = this.state;
+        return (
+            <Layout loading={loading} error={error}>
+                <LayoutContent>
+                    <LayoutTitle>Edit trigger</LayoutTitle>
+                    {trigger && (
+                        <form>
+                            <ColumnStack block gap={6} horizontalAlign="stretch">
+                                <Fit>
+                                    <ValidationContainer ref={this.validationContainer}>
+                                        {config != null && (
+                                            <TriggerEditForm
+                                                data={trigger}
+                                                tags={tags || []}
+                                                remoteAllowed={config.remoteAllowed}
+                                                onChange={update => this.handleChange(update)}
+                                            />
+                                        )}
+                                    </ValidationContainer>
+                                </Fit>
+                                <Fit>
+                                    <RowStack gap={3} baseline>
+                                        <Fit>
+                                            <Button
+                                                use="primary"
+                                                onClick={() => {
+                                                    this.handleSubmit();
+                                                }}
+                                            >
+                                                Save trigger
+                                            </Button>
+                                        </Fit>
+                                        <Fit>
+                                            <Button
+                                                use="link"
+                                                icon={<TrashIcon />}
+                                                onClick={() => {
+                                                    this.deleteTrigger(trigger.id);
+                                                }}
+                                            >
+                                                Delete
+                                            </Button>
+                                        </Fit>
+                                        <Fit>
+                                            <RouterLink to={`/trigger/${trigger.id}`}>
+                                                Cancel
+                                            </RouterLink>
+                                        </Fit>
+                                    </RowStack>
+                                </Fit>
+                            </ColumnStack>
+                        </form>
+                    )}
+                </LayoutContent>
+            </Layout>
+        );
     }
 
     async handleSubmit() {
@@ -89,6 +124,10 @@ class TriggerEditContainer extends React.Component<Props, State> {
         }
     }
 
+    handleChange(update: $Shape<Trigger>) {
+        this.setState((prevState: State) => ({ trigger: { ...prevState.trigger, ...update } }));
+    }
+
     async deleteTrigger(id: string) {
         const { history, moiraApi } = this.props;
         this.setState({ loading: true });
@@ -100,63 +139,35 @@ class TriggerEditContainer extends React.Component<Props, State> {
         }
     }
 
-    handleChange(update: $Shape<Trigger>) {
-        this.setState((prevState: State) => ({ trigger: { ...prevState.trigger, ...update } }));
+    async getData(props: Props) {
+        const { moiraApi, match } = props;
+        const { id } = match.params;
+        if (typeof id !== "string") {
+            this.setState({ error: "Wrong trigger id", loading: false });
+            return;
+        }
+        try {
+            const trigger = await moiraApi.getTrigger(id);
+            const { list } = await moiraApi.getTagList();
+            const config = await moiraApi.getConfig();
+            this.setState({
+                trigger,
+                tags: list,
+                config,
+            });
+        } catch (error) {
+            this.setState({ error: error.message });
+        } finally {
+            this.setState({ loading: false });
+        }
     }
 
-    render(): React.Node {
-        const { loading, error, trigger, tags, config } = this.state;
-        return (
-            <Layout loading={loading} error={error}>
-                <LayoutContent>
-                    <LayoutTitle>Edit trigger</LayoutTitle>
-                    {trigger && (
-                        <form>
-                            <ColumnStack block gap={6} horizontalAlign="stretch">
-                                <Fit>
-                                    <ValidationContainer ref={x => (this.triggerForm = x)}>
-                                        {config != null && (
-                                            <TriggerEditForm
-                                                data={trigger}
-                                                tags={tags || []}
-                                                remoteAllowed={config.remoteAllowed}
-                                                onChange={update => this.handleChange(update)}
-                                            />
-                                        )}
-                                    </ValidationContainer>
-                                </Fit>
-                                <Fit>
-                                    <RowStack gap={3} baseline>
-                                        <Fit>
-                                            <Button
-                                                use="primary"
-                                                onClick={() => {
-                                                    this.handleSubmit();
-                                                }}>
-                                                Save trigger
-                                            </Button>
-                                        </Fit>
-                                        <Fit>
-                                            <Button
-                                                use="link"
-                                                icon={<TrashIcon />}
-                                                onClick={() => {
-                                                    this.deleteTrigger(trigger.id);
-                                                }}>
-                                                Delete
-                                            </Button>
-                                        </Fit>
-                                        <Fit>
-                                            <RouterLink to={`/trigger/${trigger.id}`}>Cancel</RouterLink>
-                                        </Fit>
-                                    </RowStack>
-                                </Fit>
-                            </ColumnStack>
-                        </form>
-                    )}
-                </LayoutContent>
-            </Layout>
-        );
+    async validateForm(): Promise<boolean> {
+        if (this.validationContainer.current == null) {
+            return true;
+        }
+        const response = await this.validationContainer.current.validate();
+        return response;
     }
 }
