+++ src/Containers/TriggerListContainer.jsx
@@ -4,11 +4,11 @@ import queryString from "query-string";
 import { intersection, concat, difference, flattenDeep, uniq, isEqual } from "lodash";
 import moment from "moment";
 import Paging from "retail-ui/components/Paging";
+import type { ContextRouter } from "react-router-dom";
 import { getPageLink } from "../Domain/Global";
 import { withMoiraApi } from "../Api/MoiraApiInjection";
 import { getMaintenanceTime } from "../Domain/Maintenance";
 import type { Config } from "../Domain/Config";
-import type { ContextRouter } from "react-router-dom";
 import type { IMoiraApi } from "../Api/MoiraApi";
 import type { TriggerList } from "../Domain/Trigger";
 import type { Maintenance } from "../Domain/Maintenance";
@@ -37,6 +37,7 @@ type LocationSearch = {|
 
 class TriggerListContainer extends React.Component<Props, State> {
     props: Props;
+
     state: State = {
         loading: true,
         error: null,
@@ -46,53 +47,6 @@ class TriggerListContainer extends React.Component<Props, State> {
         config: null,
     };
 
-    async getData(props: Props) {
-        const { moiraApi, location } = props;
-        const { page, onlyProblems, tags: parsedTags } = this.parseLocationSearch(location.search);
-        const localDataString = localStorage.getItem("moiraSettings");
-        const { tags: localTags, onlyProblems: localOnlyProblems } =
-            typeof localDataString === "string" ? JSON.parse(localDataString) : {};
-
-        let searchToUpdate = null;
-        if (parsedTags.length === 0 && localTags && localTags.length) {
-            searchToUpdate = { ...(searchToUpdate || {}), tags: localTags };
-        }
-
-        if (!onlyProblems && localOnlyProblems) {
-            searchToUpdate = { ...(searchToUpdate || {}), onlyProblems: localOnlyProblems };
-        }
-        if (searchToUpdate != null) {
-            this.changeLocationSearch(searchToUpdate);
-            return;
-        }
-
-        try {
-            const { subscriptions } = await moiraApi.getSettings();
-            const { list: allTags } = await moiraApi.getTagList();
-            const config = await moiraApi.getConfig();
-            const selectedTags = intersection(parsedTags, allTags);
-            const triggers = await moiraApi.getTriggerList(page - 1, onlyProblems, selectedTags);
-
-            if (page > Math.ceil(triggers.total / triggers.size) && triggers.total !== 0) {
-                const rightLastPage = Math.ceil(triggers.total / triggers.size) || 1;
-                this.changeLocationSearch({ page: rightLastPage });
-                return;
-            }
-
-            this.setState({
-                config: config,
-                error: null,
-                subscriptions: uniq(flattenDeep(subscriptions.map(x => x.tags))),
-                tags: allTags,
-                triggers,
-            });
-        } catch (error) {
-            this.setState({ error: error.message });
-        } finally {
-            this.setState({ loading: false });
-        }
-    }
-
     componentDidMount() {
         this.getData(this.props);
     }
@@ -100,26 +54,20 @@ class TriggerListContainer extends React.Component<Props, State> {
     componentWillReceiveProps(nextProps: Props) {
         this.setState({ loading: true });
         this.getData(nextProps);
-        if (this.needScrollToTop(this.props, nextProps)) {
+        if (TriggerListContainer.needScrollToTop(this.props, nextProps)) {
             try {
                 window.scrollTo({
                     top: 0,
                     left: 0,
                     behavior: "smooth",
                 });
-            } catch (e) {
-                // Do nothing. Strange unstable exception on chrome
+            } catch (error) {
+                // ToDo разобраться, зачем здесь эта конструкция и удалить её
             }
         }
     }
 
-    needScrollToTop(prevProps: Props, nextProps: Props): boolean {
-        const { page: prevPage } = this.parseLocationSearch(prevProps.location.search);
-        const { page: nextPage } = this.parseLocationSearch(nextProps.location.search);
-        return !isEqual(prevPage, nextPage);
-    }
-
-    parseLocationSearch(search: string): LocationSearch {
+    static parseLocationSearch(search: string): LocationSearch {
         const { page, tags, onlyProblems } = queryString.parse(search, { arrayFormat: "index" });
         return {
             page: typeof page === "string" ? Number(page.replace(/\D/g, "")) || 1 : 1,
@@ -128,49 +76,18 @@ class TriggerListContainer extends React.Component<Props, State> {
         };
     }
 
-    changeLocationSearch(update: $Shape<LocationSearch>) {
-        const { location, history } = this.props;
-        const search = {
-            ...this.parseLocationSearch(location.search),
-            ...update,
-        };
-        localStorage.setItem("moiraSettings", JSON.stringify(search));
-        history.push(
-            "?" +
-                queryString.stringify(search, {
-                    arrayFormat: "index",
-                    encode: true,
-                })
-        );
-    }
-
-    async setMetricMaintenance(triggerId: string, maintenance: Maintenance, metric: string) {
-        this.setState({ loading: true });
-        const maintenanceTime = getMaintenanceTime(maintenance);
-        await this.props.moiraApi.setMaintenance(triggerId, {
-            metrics: {
-                [metric]:
-                    maintenanceTime > 0
-                        ? moment
-                              .utc()
-                              .add(maintenanceTime, "minutes")
-                              .unix()
-                        : maintenanceTime,
-            },
-        });
-        this.getData(this.props);
-    }
-
-    async removeMetric(triggerId: string, metric: string) {
-        this.setState({ loading: true });
-        await this.props.moiraApi.delMetric(triggerId, metric);
-        this.getData(this.props);
+    static needScrollToTop(prevProps: Props, nextProps: Props): boolean {
+        const { page: prevPage } = this.parseLocationSearch(prevProps.location.search);
+        const { page: nextPage } = this.parseLocationSearch(nextProps.location.search);
+        return !isEqual(prevPage, nextPage);
     }
 
     render(): React.Node {
         const { loading, error, triggers, tags, subscriptions, config } = this.state;
         const { location } = this.props;
-        const { page, onlyProblems, tags: parsedTags } = this.parseLocationSearch(location.search);
+        const { page, onlyProblems, tags: parsedTags } = TriggerListContainer.parseLocationSearch(
+            location.search
+        );
         const selectedTags = tags ? intersection(parsedTags, tags) : [];
         const subscribedTags = subscriptions ? difference(subscriptions, selectedTags) : [];
         const remainedTags = difference(tags, concat(selectedTags, subscribedTags));
@@ -234,13 +151,102 @@ class TriggerListContainer extends React.Component<Props, State> {
                             caption="Next page"
                             activePage={page}
                             pagesCount={pageCount}
-                            onPageChange={page => this.changeLocationSearch({ page })}
+                            onPageChange={p => this.changeLocationSearch({ page: p })}
                             withoutNavigationHint
                         />
                     </LayoutFooter>
                 )}
             </Layout>
         );
+    }
+
+    async getData(props: Props) {
+        const { moiraApi, location } = props;
+        const { page, onlyProblems, tags: parsedTags } = TriggerListContainer.parseLocationSearch(
+            location.search
+        );
+        const localDataString = localStorage.getItem("moiraSettings");
+        const { tags: localTags, onlyProblems: localOnlyProblems } =
+            typeof localDataString === "string" ? JSON.parse(localDataString) : {};
+
+        let searchToUpdate = null;
+        if (parsedTags.length === 0 && localTags && localTags.length) {
+            searchToUpdate = { ...(searchToUpdate || {}), tags: localTags };
+        }
+
+        if (!onlyProblems && localOnlyProblems) {
+            searchToUpdate = { ...(searchToUpdate || {}), onlyProblems: localOnlyProblems };
+        }
+        if (searchToUpdate != null) {
+            this.changeLocationSearch(searchToUpdate);
+            return;
+        }
+
+        try {
+            const { subscriptions } = await moiraApi.getSettings();
+            const { list: allTags } = await moiraApi.getTagList();
+            const config = await moiraApi.getConfig();
+            const selectedTags = intersection(parsedTags, allTags);
+            const triggers = await moiraApi.getTriggerList(page - 1, onlyProblems, selectedTags);
+
+            if (page > Math.ceil(triggers.total / triggers.size) && triggers.total !== 0) {
+                const rightLastPage = Math.ceil(triggers.total / triggers.size) || 1;
+                this.changeLocationSearch({ page: rightLastPage });
+                return;
+            }
+
+            this.setState({
+                config,
+                error: null,
+                subscriptions: uniq(flattenDeep(subscriptions.map(x => x.tags))),
+                tags: allTags,
+                triggers,
+            });
+        } catch (error) {
+            this.setState({ error: error.message });
+        } finally {
+            this.setState({ loading: false });
+        }
+    }
+
+    changeLocationSearch(update: $Shape<LocationSearch>) {
+        const { location, history } = this.props;
+        const search = {
+            ...TriggerListContainer.parseLocationSearch(location.search),
+            ...update,
+        };
+        localStorage.setItem("moiraSettings", JSON.stringify(search));
+        history.push(
+            `?${queryString.stringify(search, {
+                arrayFormat: "index",
+                encode: true,
+            })}`
+        );
+    }
+
+    async setMetricMaintenance(triggerId: string, maintenance: Maintenance, metric: string) {
+        const { moiraApi } = this.props;
+        this.setState({ loading: true });
+        const maintenanceTime = getMaintenanceTime(maintenance);
+        await moiraApi.setMaintenance(triggerId, {
+            metrics: {
+                [metric]:
+                    maintenanceTime > 0
+                        ? moment
+                              .utc()
+                              .add(maintenanceTime, "minutes")
+                              .unix()
+                        : maintenanceTime,
+            },
+        });
+        this.getData(this.props);
+    }
+
+    async removeMetric(triggerId: string, metric: string) {
+        const { moiraApi } = this.props;
+        this.setState({ loading: true });
+        await moiraApi.delMetric(triggerId, metric);
+        this.getData(this.props);
     }
 }
